pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-topic-mapper"
version = "0.1.0"
description = "Deterministic document structure intelligence: topic hierarchies, cross-references, entities, ambiguities."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pypdf>=4.0.0",
    "networkx>=3.0",
    "matplotlib>=3.5",
]

[project.scripts]

[tool.setuptools.packages.find]
where = ["src"]
